Moved Toggle Complete functionality to parent abstract class
