+++ FFXIVRelicTrackerBlazor/Client/Pages/AbstractSummaryBase.cs
@@ -10,7 +10,6 @@ namespace FFXIVRelicTrackerBlazor.Client.Pages
     public abstract class AbstractSummaryBase : AbstractPageBase
     {
         public List<JobName> ValidJobs => JobsByExpansion.GetJobList(TargetExpansion.Expansion);
-        public abstract AbstractExpansion TargetExpansion { get; }
         public int StageCount => TargetExpansion.StageCount;
         public List<JobName> JobNames;
         public bool showAll;
@@ -28,30 +27,13 @@ namespace FFXIVRelicTrackerBlazor.Client.Pages
             return Task.CompletedTask;
         }
 
-        public bool GetAdjustCount => TargetExpansion.AdjustCounts;
         public string AdjustCountString => TargetExpansion.AdjustCounts ? "will" : "will not";
         public async Task ToggleAdjustCount()
         {
             TargetExpansion.AdjustCounts = !TargetExpansion.AdjustCounts;
             await OnCharacterUpdate();
         }
-        public async Task ToggleComplete(int StageIndex, JobName job)
-        {
-            if (TargetExpansion.Jobs.Where(x => x.JobName == job).First().Stages.Where(x => x.StageIndex == StageIndex).First().Progress == Progress.Completed)
-                await InCompleteStage(StageIndex, job);
-            else await CompleteStage(StageIndex, job);
-
-        }
-        private async Task CompleteStage(int StageIndex, JobName job)
-        {
-            MasterStageHelper.CompleteStage(character, job, StageIndex, TargetExpansion.Expansion, GetAdjustCount);
-            await OnCharacterUpdate();
-        }
-        private async Task InCompleteStage(int StageIndex, JobName job)
-        {
-            MasterStageHelper.InCompleteStage(character, job, StageIndex, TargetExpansion.Expansion);
-            await OnCharacterUpdate();
-        }
+        
 
     }
 }
